+++ src/Command/GenerateCommand.php
@@ -8,13 +8,10 @@ use Helmich\Schema2Class\Generator\GeneratorException;
 use Helmich\Schema2Class\Generator\GeneratorRequest;
 use Helmich\Schema2Class\Generator\GenerationRunner;
 use Helmich\Schema2Class\Loader\LoadingException;
-use Helmich\Schema2Class\Loader\SchemaLoader;
+use Helmich\Schema2Class\Spec\Specification;
+use Helmich\Schema2Class\Spec\SpecificationFilesItem;
 use Helmich\Schema2Class\Spec\SpecificationOptions;
 use Helmich\Schema2Class\Spec\OptionsDefaults;
-use Helmich\Schema2Class\Spec\ValidatedSpecificationFilesItem;
-use Helmich\Schema2Class\Util\StringUtils;
-use Helmich\Schema2Class\Generator\Property\IntersectProperty;
-use Helmich\Schema2Class\Generator\Property\NestedObjectProperty;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -25,14 +22,12 @@ class GenerateCommand extends Command
 {
 
     private GenerationRunner $runner;
-    private SchemaLoader $loader;
 
-    public function __construct(GenerationRunner $runner, SchemaLoader $loader)
+    public function __construct(GenerationRunner $runner)
     {
         parent::__construct();
 
         $this->runner = $runner;
-        $this->loader = $loader;
     }
 
     protected function configure(): void
@@ -81,37 +76,19 @@ class GenerateCommand extends Command
         /** @var string|null $targetPHPVersion */
         $targetPHPVersion = $input->getOption("target-php");
 
-        $output->writeln("loading schema from <comment>$schemaFile</comment>");
-        $schema = $this->loader->loadSchema($schemaFile);
-        $needsClass = IntersectProperty::canHandleSchema($schema)
-            || NestedObjectProperty::canHandleSchema($schema)
-            || array_key_exists('enum', $schema);
-        if ($class === null && $needsClass) {
-            $basename = pathinfo($schemaFile, PATHINFO_FILENAME);
-            $class = StringUtils::pascalCase($basename);
+        $opts = new SpecificationOptions();
+        $opts = $opts->withTargetDirectory($targetDirectory);
+        if ($targetNamespace) {
+            $opts = $opts->withTargetNamespace($targetNamespace);
         }
-
-        $targetNamespace = $this->runner->inferNamespace($targetDirectory, $targetNamespace, $output);
-
-        $output->writeln("using target namespace <comment>$targetNamespace</comment> in directory <comment>$targetDirectory</comment>");
-
-        $cleanTarget = (bool)$input->getOption('clean-dir');
-
-        if ($cleanTarget) {
-            $this->runner->cleanDirectory($targetDirectory, $output);
-        }
-
-        $spec = new ValidatedSpecificationFilesItem($targetNamespace, $class, $targetDirectory, $cleanTarget);
+        $opts = $opts->withCleanTargetDirectory((bool)$input->getOption('clean-dir'));
 
         if (!$targetPHPVersion) {
             $targetPHPVersion = GeneratorRequest::DEFAULT_PHP8_VERSION;
         } else {
             $targetPHPVersion = GeneratorRequest::normalizeTargetVersion($targetPHPVersion);
         }
-
-        $opts = (new SpecificationOptions())->withTargetPHPVersion($targetPHPVersion);
-
-        $output->writeln("target PHP version: <comment>{$opts->getTargetPHPVersion()}</comment>");
+        $opts = $opts->withTargetPHPVersion($targetPHPVersion);
 
         if ($input->getOption("disable-strict-types")) {
             $opts = $opts->withDisableStrictTypes(true);
@@ -146,9 +123,14 @@ class GenerateCommand extends Command
 
         $opts = OptionsDefaults::applyDefaults($opts);
 
-        $baseRequest = new GeneratorRequest($schema, $spec, $opts);
+        $file = new SpecificationFilesItem($schemaFile);
+        if ($class !== null) {
+            $file = $file->withClassName($class);
+        }
 
-        $this->runner->generateFromRequest($baseRequest, $output, (bool)$input->getOption('dry-run'));
+        $spec = (new Specification([$file]))->withOptions($opts);
+
+        $this->runner->generateFromSpecification($spec, $output, (bool)$input->getOption('dry-run'));
 
         return 0;
     }